pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools_scm]
version_file = "route_planner/route_planner/__init__.py"

[project]
name = "route-planner"
description="Route Planner"
dynamic = ["version"]
authors = [
    {name = "Michal Mládek", email = "osvc.04923031@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
keywords = ["maps", "internet", "traffic", "route"]
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "python-dotenv",
    #"python-dateutil",
    "django",
    "django-ninja",
    "django-extensions",
    #"django-filter",
    "djangorestframework",
    "folium",
    "openrouteservice",
]

[project.optional-dependencies]
dev = [
    "ipython",
    "flake8",
    "isort",
    #"pytest",
    #"pytest-django",
    "mypy",
    "pre-commit",
    #"faker",
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    #"markdown",
    #"rich"
]
prod = ["gunicorn[gevent]"]

[project.urls]
homepage = "https://route-planner.softorks.eu"
repository = "https://github.com/michalnik/route-planner"
